♻️ Refactor: Table components
