+++ client/src/ui/Table/Table.module.scss
@@ -1,5 +1,6 @@
 .table {
-  @apply w-full text-sm text-neutral-500 border-spacing-0;
+  @apply w-full;
+  @apply text-sm text-neutral-500 border-spacing-0;
 
   td {
     @apply p-2;
@@ -9,39 +10,17 @@
     //}
 
     &:first-child {
-      @apply pl-4 rounded-tl rounded-bl;
+      @apply pl-4;
+      @apply rounded-tl rounded-bl;
     }
 
     &:last-child {
-      @apply pr-4 rounded-tr rounded-br;
+      @apply pr-4;
+      @apply rounded-tr rounded-br;
     }
   }
 
   tr:hover td {
     @apply bg-white bg-opacity-10;
   }
-}
-
-.tableHeader {
-  @apply border-b border-neutral-500 border-opacity-20;
-
-  td span {
-    @apply relative inline-flex items-center select-none cursor-pointer;
-
-    svg {
-      @apply absolute -right-6 text-green-500 scale-150;
-    }
-  }
-
-  td span:hover {
-    @apply text-white;
-  }
-}
-
-// Hack: space between thead and tbody
-.tableBody:before {
-  content: '@';
-  display: block;
-  line-height: 8px;
-  text-indent: -99999px;
 }
