+++ client/src/ui/Table/TableBody.module.scss
@@ -1,0 +1,7 @@
+// Hack: space between thead and tbody
+.tableBody:before {
+  content: '@';
+  display: block;
+  line-height: 8px;
+  text-indent: -99999px;
+}
