+++ client/src/ui/Table/TableRow.tsx
@@ -1,4 +1,3 @@
-import styles from './Table.module.scss';
 import { ReactNode } from 'react';
 
 interface TableRowProps {
@@ -6,7 +5,7 @@ interface TableRowProps {
 }
 
 const TableRow = ({ children }: TableRowProps) => {
-  return <tr className={styles.tableRow}>{children}</tr>;
+  return <tr>{children}</tr>;
 };
 
 export default TableRow;
